backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
pymysql==1.1.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
zhipuai==2.1.5.20250801
python-docx==1.1.2
pymupdf==1.25.5
httpx==0.27.0
jinja2==3.1.4
numpy==2.2.6
scikit-learn==1.6.1
python-dotenv==1.0.1
